# docs/api/api.yaml
openapi: "3.0.0"
info:
  title: "Wutsi Security Server"
  description: |
    Wutsi Security Server. It provides the following features:

    - [JWT](https://jwt.io) Authentication
    - Application registration
  version: "1.0.0"
servers:
  - url: https://wutsi-security-server-test.herokuapp.com
    description: Sandbox
  - url: https://wutsi-security-server-prod.herokuapp.com
    description: Production
paths:
...
  /v1/applications/{id}/scopes:
    post:
      operationId: "grant-scopes"
      description: |
        Grant a scope to an application
      parameters:
        - in: path
          name: id
          description: ID of the application
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GrantScopeRequest'
      responses:
        200:
          description: Success
      tags:
        - Application
...
components:
  schemas:
...
    GrantScopeRequest:
      type: object
      properties:
        scopeNames:
          type: array
          items:
            type: string
      required:
        - scopeNames
...